pyproject.toml
[project]
name = "rl-gym"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.0.0",
    "ipykernel>=6.30.0",
    "notebook>=7.4.4",
    "openai>=1.96.1",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
    "verifiers>=0.0.0",
    "weave>=0.51.56",
]
